refactor(SR-8): separate createApp from portal server index
